add parser folder to annotation
